python/dashy/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --allow-unsafe requirements.in
#
ansi2html==1.8.0
    # via jupyter-dash
appnope==0.1.3
    # via
    #   ipykernel
    #   ipython
asttokens==2.0.8
    # via stack-data
backcall==0.2.0
    # via ipython
brotli==1.0.9
    # via flask-compress
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
dash==2.6.2
    # via
    #   -r requirements.in
    #   dash-bootstrap-components
    #   jupyter-dash
dash-bootstrap-components==1.2.1
    # via -r requirements.in
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
debugpy==1.6.3
    # via ipykernel
decorator==5.1.1
    # via ipython
duckdb==0.5.1
    # via -r requirements.in
entrypoints==0.4
    # via jupyter-client
executing==1.1.0
    # via stack-data
flask==2.2.2
    # via
    #   dash
    #   flask-compress
    #   jupyter-dash
flask-compress==1.13
    # via dash
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
ipykernel==6.16.0
    # via jupyter-dash
ipython==8.5.0
    # via
    #   ipykernel
    #   jupyter-dash
itsdangerous==2.1.2
    # via flask
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via flask
jupyter-client==7.3.5
    # via ipykernel
jupyter-core==4.11.1
    # via jupyter-client
jupyter-dash==0.4.2
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-dash
numpy==1.23.3
    # via
    #   duckdb
    #   pandas
packaging==21.3
    # via ipykernel
pandas==1.5.0
    # via -r requirements.in
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
plotly==5.10.0
    # via dash
prompt-toolkit==3.0.31
    # via ipython
psutil==5.9.2
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.13.0
    # via ipython
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.4
    # via pandas
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
requests==2.28.1
    # via jupyter-dash
retrying==1.3.3
    # via jupyter-dash
six==1.16.0
    # via
    #   python-dateutil
    #   retrying
stack-data==0.5.1
    # via ipython
tenacity==8.1.0
    # via plotly
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
traitlets==5.4.0
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   matplotlib-inline
urllib3==1.26.12
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.2.2
    # via flask

# The following packages are considered to be unsafe in a requirements file:
setuptools==65.4.1
    # via gunicorn
